+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 # Changelog
 
 ## Unreleased changes
-- CredentialRegistrationID type added and AccountAddress can be derived from it.
+
+## 2.1.0
+- CredentialRegistrationID type introduced and AccountAddress can be derived from it.
 
 ## 2.0.0
 - Support for configuring TLS.